refactor: 컴포넌트 폴더 구조를 common과 layout으로 정리
