+++ docs/COMMON_COMPONENT_GUIDE.md
@@ -154,18 +154,31 @@
 처음부터 모든 공통 컴포넌트를 만들면 오히려 느려진다. 아래 순서로 먼저 만든다.
 
 1. `AppShell`
+   - 화면 전체의 기본 틀이다. 좌측 사이드바, 상단 헤더, 본문 영역 배치를 공통으로 잡는다.
 2. `SideNav`
+   - 서비스 메뉴를 보여주는 좌측 네비게이션이다.
 3. `TopHeader`
+   - 상단 검색, 알림, 사용자 정보 영역이다.
 4. `BaseButton`
+   - 저장, 취소, 등록, 수정, 삭제처럼 반복되는 버튼 공통 컴포넌트다.
 5. `BaseTextField`
+   - 일반 텍스트 입력창이다. 이름, 제목, 코드, 이메일 입력 등에 사용한다.
 6. `BaseSelect`
+   - 상태, 역할, 부서, 거래처 같은 선택형 입력에 사용한다.
 7. `FormField`
+   - 라벨, 필수 표시, 입력 컴포넌트를 한 묶음으로 보여주는 폼 단위 컴포넌트다.
 8. `SearchInput`
+   - 목록 조회나 상단 검색에 사용하는 검색 전용 입력창이다.
 9. `StatusBadge`
+   - 활성, 비활성, 확정, 초안, 발송, 완료 같은 상태를 공통 스타일로 보여준다.
 10. `BaseTable`
+   - 목록 화면의 표 구조를 공통으로 처리하는 컴포넌트다.
 11. `SectionCard`
+   - 상세 화면에서 제목이 있는 박스형 섹션을 공통으로 처리한다.
 12. `InfoField`
+   - 라벨과 값을 한 세트로 보여주는 상세 정보 표시 컴포넌트다.
 13. `BaseModal`
+   - 등록, 수정, 상세 보기 팝업의 기본 틀을 담당한다.
 
 이 1차 세트만 있어도 아래 화면 대부분을 시작할 수 있다.
 
@@ -180,12 +193,19 @@
 1차 공통이 안정되면 다음을 만든다.
 
 1. `LineItemTable`
+   - 문서 화면에서 반복되는 품목 / 수량 / 단가 / 금액 테이블이다.
 2. `DocumentHeaderActions`
+   - 문서 상세 상단의 수정, 삭제, 미리보기, 인쇄, PDF 다운로드 버튼 묶음이다.
 3. `DocumentSummarySection`
+   - PI, PO, CI, PL 같은 문서의 기본 정보를 보여주는 요약 영역이다.
 4. `LinkedDocumentList`
+   - 문서 간 연결 관계를 보여주는 목록 컴포넌트다.
 5. `FilterPanel`
+   - 조회 조건, 기간, 드롭다운, 검색/초기화 버튼을 묶은 필터 영역이다.
 6. `ActivityTypeBadge`
+   - 활동 유형을 공통 색상과 형태로 보여주는 배지 컴포넌트다.
 7. `ActivityDetailModal`
+   - 활동 기록 상세 내용을 팝업 형태로 보여주는 컴포넌트다.
 
 이 단계부터는 문서 화면과 활동 기록 화면 구현 속도가 빨라진다.
 
@@ -301,4 +321,3 @@ src/components/
 - 공통 컴포넌트는 `여러 화면에서 반복되는 단순 UI`만 먼저 만든다.
 - `문서 전용`, `활동 전용`, `인코텀즈 전용` 같은 것은 도메인 컴포넌트로 분리한다.
 - 먼저 작은 공통 세트를 고정하고, 그 위에서 각 서비스 화면을 조립하는 방식으로 가야 팀이 가장 이해하기 쉽고 충돌도 적다.
-
