+++ src/layouts/AppLayout.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import AppHeader from '@/components/AppHeader.vue'
-import AppSidebar from '@/components/AppSidebar.vue'
+import AppHeader from '@/components/layout/AppHeader.vue'
+import AppSidebar from '@/components/layout/AppSidebar.vue'
 </script>
 
 <template>
