+++ tsconfig.json
@@ -4,7 +4,7 @@
     "useDefineForClassFields": true,
     "module": "ESNext",
     "moduleResolution": "Node",
-    "baseUrl": "./frontend",
+    "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"]
     },
